[Fix] 메인 뷰 레이아웃 수정 (#99)

+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/Main/Models/MainModel.swift
@@ -41,7 +41,7 @@ struct DateCourseModel {
     let city: String
     let like: Int
     let cost: Int
-    let duration: Int
+    let duration: Float
     
     static var hotDateDummyData: [DateCourseModel] {
         return [DateCourseModel(courseId: 1, thumbnail: "www.asdfasdfds.jpg", title: "스리와 데이트 - 성수편 스리와 데이트 - 성수편", city: "건대/상수/왕십리", like: 3, cost: 10, duration: 10),
